refactor code to ask user and find tile
