+++ src/board/board.ts
@@ -24,3 +24,19 @@ export const renderBoard = (data: Board) => {
   });
   console.log(chalk.white.bold("-------------"));
 };
+
+export const findTile = (board: Board, toFind: Tile) => {
+  let tile: Tile = null;
+  board.forEach((line) => {
+    if (tile != undefined) {
+      return;
+    }
+
+    tile = line.find((tile: any) => {
+      if (tile != undefined)
+        return tile.symbol === toFind.symbol && tile.color === toFind.color;
+    });
+  });
+
+  return tile;
+};
